+++ NEWS.md
@@ -20,6 +20,8 @@
 
 * Added `rmd_fenced_div_wrap()` function to wrap selected nodes with fenced divs
 
+* Added `rmd_insert()` function to insert nodes relative to selected nodes
+
 * Added support for code block literals - e.g. ```{{r}}
 
 * Added support for 2ndary parsing to handle inline elements: